+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.11.4] - 2026-06-09
+
 ### Fixed
 
 - `controls_from` mirrors with nested variants now hide the inactive nested